added concatenate function and axes to both stack and concatenate

+++ cardio_rl/utils.py
@@ -5,12 +5,12 @@ import jax.numpy as jnp
 # TODO: rename module so as to not be confused with typing
 
 
-def to_np(arr: np.ndarray):
-    return arr.astype(float)
+def to_np(arr: np.ndarray, dtype=np.float32):
+    return arr.astype(dtype)
 
 
-def to_jnp(arr: np.ndarray):
-    return jnp.asarray(arr, dtype=jnp.float32)
+def to_jnp(arr: np.ndarray, dtype=jnp.float32):
+    return jnp.asarray(arr, dtype=dtype)
 
 
 def to_torch(arr: np.ndarray):
